muyaho-api/src/main/java/com/depromeet/muyaho/service/member/MemberService.java
package com.depromeet.muyaho.service.member;

import com.depromeet.muyaho.domain.member.MemberRepository;
import com.depromeet.muyaho.service.member.dto.request.SignUpMemberRequest;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@RequiredArgsConstructor
@Service
public class MemberService {

    private final MemberRepository memberRepository;

    @Transactional
    public Long signUpMember(SignUpMemberRequest request) {
        return memberRepository.save(request.toEntity()).getId();
    }

}
